requirements.txt
attrs==23.1.0
blinker==1.6.2
cattrs==23.1.1
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
platformdirs==3.5.1
psycopg2==2.9.6
requests==2.28.1
requests-cache==1.0.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Unidecode==1.3.6
url-normalize==1.4.3
urllib3==1.26.16
Werkzeug==2.3.4

XlsxWriter~=3.1.2
celery~=5.3.0
PyJWT~=2.7.0